chore: improve signalStore tests
